backend/pyproject.toml
[project]
name = "zhengcheng-fastapi-hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.109.2",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.1.0",
    "sqlmodel>=0.0.24",
    "uvicorn==0.27.1",
]
